BaytesPMCMC

Documentation for BaytesPMCMC.

BaytesPMCMC.PMCMC

BaytesPMCMC.PMCMCConstructor

BaytesPMCMC.PMCMCDefault

BaytesPMCMC.PMCMCDiagnostics

BaytesPMCMC.PMCMCTune

BaytesPMCMC.ParticleGibbs

BaytesPMCMC.ParticleMetropolis

BaytesCore.generate_showvalues

BaytesCore.infer

BaytesCore.infer

BaytesCore.propose!

BaytesCore.propose!

BaytesCore.propose!

BaytesPMCMC.PMCMC — Type

struct PMCMC{M<:PMCMCKernel, N<:PMCMCTune} <: BaytesCore.AbstractAlgorithm

Particle MCMC container, container kernel and tuning.

Fields

kernel::PMCMCKernel

PMCMC sampler

tune::PMCMCTune

Tuning configuration for kernel.

BaytesPMCMC.PMCMCConstructor — Type

Callable struct to make initializing PMCMC sampler easier in sampling library.

Examples

BaytesPMCMC.PMCMCDefault — Type

struct PMCMCDefault

Default arguments for PMCMC constructor.

Fields

BaytesPMCMC.PMCMCDiagnostics — Type

struct PMCMCDiagnostics{T, P<:BaytesFilters.ParticleFilterDiagnostics, M<:BaytesMCMC.MCMCDiagnostics} <: BaytesCore.AbstractDiagnostics

Contains information about log-likelihood, expected sample size and proposal trajectory.

Fields

base::BaytesCore.BaseDiagnostics

Diagnostics used for all Baytes kernels

pf::BaytesFilters.ParticleFilterDiagnostics

Particle Filter diagnostics.

mcmc::BaytesMCMC.MCMCDiagnostics

MCMC diagnostics.

BaytesPMCMC.PMCMCTune — Type

struct PMCMCTune{T<:ModelWrappers.Tagged} <: BaytesCore.AbstractTune

PMCMC tuning container.

Fields

tagged::ModelWrappers.Tagged

Tagged Model parameter.

BaytesPMCMC.ParticleGibbs — Type

struct ParticleGibbs{P<:BaytesFilters.ParticleFilter, M<:BaytesMCMC.MCMC} <: PMCMCKernel

Default arguments for PMCMC constructor.

Fields

pf::BaytesFilters.ParticleFilter

Particle Filter kernel to estimate latent trajectory.

mcmc::BaytesMCMC.MCMC

MCMC kernel to sample continuous model parameter.

BaytesPMCMC.ParticleMetropolis — Type

struct ParticleMetropolis{P<:BaytesFilters.ParticleFilter, M<:BaytesMCMC.MCMC} <: PMCMCKernel

Default arguments for PMCMC constructor.

Fields

pf::BaytesFilters.ParticleFilter

Particle Filter kernel to estimate latent trajectory.

mcmc::BaytesMCMC.MCMC

MCMC kernel to sample continuous model parameter.

BaytesCore.generate_showvalues — Method

generate_showvalues(diagnostics)

Show relevant diagnostic results.

Examples

BaytesCore.infer — Method

infer(_rng, pmcmc, model, data)

Infer type of predictions of kernel.

Examples

BaytesCore.infer — Method

infer(_rng, diagnostics, pmcmc, model, data)

Infer PMCMC diagnostics type.

Examples

BaytesCore.propose! — Method

propose!(_rng, pmcmc, model, data)
propose!(_rng, pmcmc, model, data, temperature)
propose!(_rng, pmcmc, model, data, temperature, update)

Propose new parameter with pmcmc sampler. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.propose! — Method

propose!(_rng, pmcmc, model, data)
propose!(_rng, pmcmc, model, data, temperature)
propose!(_rng, pmcmc, model, data, temperature, update)

Propose new parameter with mcmc psampler. If update=true, objective function will be updated with input model and data.

Examples

BaytesCore.propose! — Method

propose!(_rng, pmcmc, model, data)
propose!(_rng, pmcmc, model, data, temperature)
propose!(_rng, pmcmc, model, data, temperature, update)

Propose new parameter with pmcmc sampler. If update=true, objective function will be updated with input model and data.

Examples